simrs-aklik-new/requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
git+https://github.com/jeagercoder/bpjs_lib.git
certifi==2023.5.7
charset-normalizer==3.1.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.0.1
django-cors-headers==4.0.0
django-import-export==3.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
future==0.18.3
humanize==4.6.0
idna==3.4
ipdb==0.13.13
ipython==8.13.2
jedi==0.18.2
lzstring==1.0.4
MarkupPy==1.14
matplotlib-inline==0.1.6
odfpy==1.4.1
openpyxl==3.1.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.18.0
Pygments==2.15.1
pytz==2023.3
PyYAML==6.0
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tablib==3.4.0
tomli==2.0.1
traitlets==5.9.0
urllib3==2.0.2
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
